pyproject.toml
[project]
name = "m03w01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "dbt-core>=1.10.4",
    "dbt-snowflake>=1.10.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-litellm>=0.2.1",
    "langgraph>=0.4.8",
    "litellm>=1.72.4",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
    "reportlab>=4.4.2",
    "rich>=14.0.0",
    "scikit-learn>=1.7.0",
    "setuptools>=80.9.0",
    "snowflake-connector-python>=3.16.0",
    "streamlit>=1.45.1",
    "watchdog>=6.0.0",
]
